feat: finished support for various notification messages for channel events. Refactored some settings to be placed in the new 'notifications' category to make it a bit better manageable in the side-panel.

+++ src/main/java/com/twitchliveloadout/marketplace/products/TwitchProductCost.java
@@ -1,6 +1,6 @@
 package com.twitchliveloadout.marketplace.products;
 
 public class TwitchProductCost {
-	public Integer amount = 0;
+	public Double amount = 0d;
 	public String type;
 }
